requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-prettyjson==0.4.1
idna==3.10
psycopg2-binary==2.9.10
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
standardjson==0.3.1
tzdata==2025.2
urllib3==2.5.0
pydantic~=2.12.5
python-dateutil~=2.9.0
# Email fetch (as in is_product)
django-mailbox

